Test UI (#41)
* Add tests page

* Add shadcn dialog

* Add initial code for modal to create a test

* Make listing tests public

* Adding Calendar and check boxes to the list of students

* Added test form. Need to change what is being used to set dates of the test opening and test closing

* Fixed Sign up page to have a blank tablehead if no boxes are checked and updated the date selection when creating tests.

* Test Name

Added the name of the test in the table next to the course code

* Cleaning up test file and worked on the calender not filtering properly

* Fix part 2 of the filtering issue

* Create/Edit Test

Logged in users can create/edit tests

* Updated tests

Create/Edit tests broke but they are good now

* Fixed test open and close dates showing up wrong on test cards

* Fixing times for test open and close not being auto set to correct time.

* Fixing Sorting function on signup page

* Added rule column to db and added the ability to search by student name

* Removed package-lock.json, removed comments that aren't needed, and removed functon generateId

* Testing for Erin

* Attempt Fix on Searching by Names

* Comment on Tests.tsx

added comment for angry piece of code.

* Readding SignUp.tsx to before the attempt to fix conflict

* Fixing bugs

* Enrollment

added a field for profs to put how many student will be enrolled. have to add a max_enrollments column to the tests table

* Added amount of students that have signed up for the test

* Filtering

Switched up how the filtering and pages work

* Fixing Sign up and Tests

* Fixed issue of test not being formatted correctly and enrollments not being used

* Update fields layout

put requested fields next to each other

* Fixing bug for student counts

Using the new code it will now properly show the amount of students that have signed up. There also is no longer the issue of the auto cancel.

* Revisions

Updated the code as per requested in previous comments

* Resolving reqested changes

* Added max_enrollments to my database again

* Adding test as generic

* Remove generic

* Remove executable

---------

Co-authored-by: Richard Grover <38992201+richgrov@users.noreply.github.com>
Co-authored-by: Madison Pantheon <madison@webdevbydesign.com>